+++ src/Place.Interfaces/Models/PixelSlice.cs
@@ -3,6 +3,6 @@
     public class PixelSlice
     {
         public Pixel[] Pixels { get; set; }
-        public short Offset { get; set; }
+        public int Offset { get; set; }
     }
 }
